requirements.txt
python-jsonrpc
pika
pymongo
requests
lxml
xmljson
numpy
pandas
IPython
watchdog
PySocks